experiment/requirements.txt
strands-agents[bidi]
botocore[crt]
